Esqueleto

+++ stores-service/src/main/java/cl/utalca/utaleats/storesservice/service/StoreService.java
@@ -1,14 +1,14 @@
 package cl.utalca.utaleats.storesservice.service;
 
+import cl.utalca.utaleats.storesservice.exception.ResourceNotFoundException;
 import cl.utalca.utaleats.storesservice.model.Store;
 import cl.utalca.utaleats.storesservice.repository.StoreRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
-public class StoreService implements IStoreService{
+public class StoreService implements IStoreService {
 
     private final StoreRepository storeRepository;
 
@@ -16,23 +16,28 @@ public class StoreService implements IStoreService{
         this.storeRepository = storeRepository;
     }
 
+    @Override
     public List<Store> getStoresByCity(String city) {
         return storeRepository.findByCity(city);
     }
 
+    @Override
     public List<Store> getAllStores() {
         return storeRepository.findAll();
     }
 
+    @Override
     public Store getStoreById(Long id) {
-        Optional<Store> optionalStore = storeRepository.findById(id);
-        return optionalStore.orElse(null); // o lanza una excepción si prefieres
+        return storeRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Tienda con ID " + id + " no encontrada"));
     }
 
+    @Override
     public Store createStore(Store store) {
         return storeRepository.save(store);
     }
 
+    @Override
     public Store updateStore(Long id, Store updatedStore) {
         return storeRepository.findById(id)
                 .map(store -> {
@@ -43,13 +48,14 @@ public class StoreService implements IStoreService{
                     store.setRating(updatedStore.getRating());
                     return storeRepository.save(store);
                 })
-                .orElseGet(() -> {
-                    updatedStore.setId(id);
-                    return storeRepository.save(updatedStore);
-                });
+                .orElseThrow(() -> new ResourceNotFoundException("Tienda con ID " + id + " no encontrada"));
     }
 
+    @Override
     public void deleteStore(Long id) {
+        if (!storeRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Tienda con ID " + id + " no encontrada");
+        }
         storeRepository.deleteById(id);
     }
 }
